requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
appdirs=1.4.4=pypi_0
blas=1.0=mkl
braindecode=0.4.85=pypi_0
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2022.3.29=haa95532_0
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.15.0=py310h2bbff1b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.4=pypi_0
cryptography=36.0.0=py310h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
fonttools=4.32.0=pypi_0
freetype=2.10.4=hd328e21_0
future=0.18.2=pypi_0
h5py=3.6.0=pypi_0
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
jinja2=3.1.1=pypi_0
joblib=1.1.0=pypi_0
jpeg=9d=h2bbff1b_0
kiwisolver=1.4.2=pypi_0
libffi=3.4.2=h604cdb4_1
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
llvmlite=0.38.0=pypi_0
lz4-c=1.9.3=h2bbff1b_1
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
mne=1.0.1=pypi_0
numba=0.55.1=pypi_0
numpy=1.21.5=py310h4c31df0_0
numpy-base=1.21.5=py310hedd7904_0
openssl=1.1.1n=h2bbff1b_0
packaging=21.3=pypi_0
pandas=1.4.2=pypi_0
pillow=9.0.1=py310hdc2b20a_0
pip=21.2.4=py310haa95532_0
pooch=1.6.0=pypi_0
psutil=5.9.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pymatreader=0.0.30=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.8=pypi_0
pyprep=0.4.2=pypi_0
pysocks=1.7.1=py310haa95532_0
python=3.10.4=hbb2ffb3_0
python-dateutil=2.8.2=pypi_0
pytorch=1.11.0=py3.10_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pypi_0
requests=2.27.1=pyhd3eb1b0_0
resampy=0.2.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
setuptools=58.0.4=py310haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.2=h2bbff1b_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h2bbff1b_0
torchaudio=0.11.0=py310_cu113
torchvision=0.12.0=py310_cu113
tqdm=4.64.0=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.8=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py310haa95532_0
wincertstore=0.2=py310haa95532_2
xmltodict=0.12.0=pypi_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=hbd8134f_5
zstd=1.4.9=h19a0ad4_0
